refactor: LoginPage component
